+++ analysis/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ## [unreleased] (Added 🚀 | Changed | Removed  | Fixed 🐞 | Chore 👨‍💻 👩‍💻)
 
+### Added 🚀
+
+- Add Delphi support to UnifiedParser (.pas, .dpr)
+
 ## [1.142.0] - 2026-03-16
 
 ### Added 🚀